Update cmake req version and version handling (#74)
The version is now calculated from the current git tag. Matching the
pattern vX.Y.Z. The latest patch in the branch will be used in the
build.

This change required the CI action to do a non-shallow checkout. Which
should be fine.

Also removed the 'dry-run' flag from the steam build vdf.

Installs cmocka to enable test-runs in CI.