Improve log rendering performance
Logs are now stored in a circular buffer rather then a shifting list of
strings. This will improve the performance when adding logs but should
not have that much impace since we don't do this on every frame.

However the rendering of the logs has been improved to reduce the
performance impact when generating textures for each log line on every
frame.